First Commit
Folders updated:
Blink: made it so an LED toggles when a button is pressed.
gpio: changed the 7 segment display so that it displays only a single segment on command.
pwm/measure_duty_cycle: changed the original code so that it will gradually incease and decrease the brigtness of an LED.